Updated WebGL 2.0 renderer to use uniform buffers
Uses uniform buffer for all uniforms that are unchanged for the
duration of the frame. Need to convert per-material uniforms to uniform
buffers as well.

Also updated how we use stats.js (and tinkered with it's guts a bit) to
better show the thing we really care about here, which is how much time
we spend in javascript submitting each frame. (It's generally assumed
that the GPU-side work is ~= identical between renderers.)